system-test: fix order in cleanup function
Firewalls must be deleted before networks. Only call the callback
when all deletes have finished or if an error occurred.
Otherwise the deletes are not happening in the correct order.

async.forEach runs only a single async operation at a
time, but the callback is called after
every single deletion and therefore too early.